+++ src/test/java/samTextViewer/TrackProcessorTest.java
@@ -40,30 +40,4 @@ public class TrackProcessorTest {
 		TrackProcessor tp= new TrackProcessor(trackSet, gch); 
 		// tp.iterateTracks();
 	}
-	
-//	@Test
-//	public void canSaveTrackSettings() throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, BamIndexNotFoundException, InvalidRecordException, SQLException{
-//		
-//		GenomicCoords gc= new GenomicCoords("chr7:1-100", null, null);
-//		List<String> genome= new ArrayList<String>();
-//		genome.add("test_data/ear045.oxBS.actb.bam");
-//		gc.setGenome(genome, true);
-//
-//		GenomicCoordsHistory gch= new GenomicCoordsHistory();
-//		gch.add(gc);
-//		
-//		TrackSet trackSet= new TrackSet();
-//		trackSet.addTrackFromSource("test_data/hg19_genes.gtf.gz", gc, null);
-//		trackSet.addTrackFromSource("test_data/ear045.oxBS.actb.bam", gc, null);
-//		trackSet.addTrackFromSource("test_data/ear045.oxBS.actb.tdf", gc, null);
-//		
-//		TrackProcessor tp= new TrackProcessor(trackSet, gch); 
-//		
-//		File x= new File("deleteme.txt");
-//		x.deleteOnExit();
-//		tp.exportTrackSetSettings(x.getAbsolutePath());
-//		assertTrue(x.length() > 200); // File not empty.
-//		
-//	}
-
 }